requirements.txt
python-docx>=1.0.0
pypdf>=4.0.0
openai>=1.0.0
